rename scala.meta.internal.converters
This package name is not specific enough. I’d like to also include the name
of the plugin to say `scala.meta.internal.scalahost.converters`.
