+++ scalahost/src/main/scala/scala/meta/internal/scalahost/converters/Converter.scala
@@ -1,5 +1,6 @@
 package scala.meta
 package internal
+package scalahost
 package converters
 
 trait Converter extends ReflectToolkit with ToMtree
