+++ scalahost/src/main/scala/scala/meta/internal/scalahost/converters/ReflectToolkit.scala
@@ -1,5 +1,6 @@
 package scala.meta
 package internal
+package scalahost
 package converters
 
 import scala.collection.mutable
