+++ scalahost/src/test/scala/scala/meta/tests/scalahost/ConverterSuite.scala
@@ -7,7 +7,7 @@ import scala.tools.cmd.CommandLineParser
 import scala.tools.nsc.{Global, CompilerCommand, Settings}
 import scala.tools.nsc.reporters.StoreReporter
 import org.scalatest._
-import scala.meta.internal.converters.Converter
+import scala.meta.internal.scalahost.converters.Converter
 
 trait ConverterSuite extends FunSuiteLike {
