+++ scalahost/src/test/scala/scala/meta/tests/scalahost/LotsOfProjects.scala
@@ -2,7 +2,7 @@ package scala.meta.tests
 package scalahost
 
 import scala.collection.mutable
-import scala.meta.internal.converters.ConvertException
+import scala.meta.internal.scalahost.converters.ConvertException
 import scala.meta.testkit._
 import scala.util.Try
